chore: Refactor mana calculator (#1315)
* feat: Flexible network selector in the mana calculator

* feat: Improved user flow for the Mana calculator

* typo

* chore: Refactor the Mana Calculator

* :sparkles:

* clean up

* clean up

* fmt

* fix

* fix

---------
