+++ src/components/ManaCalculator/enums/index.ts
@@ -1,0 +1,1 @@
+export * from './parameters.enum';
