+++ src/components/ManaCalculator/index.ts
@@ -1,0 +1,1 @@
+export { ManaCalculator as default } from './components/ManaCalculator';
