+++ src/components/ManaCalculator/types/index.ts
@@ -1,0 +1,2 @@
+export * from './mana-calculator.type';
+export * from './validator.type';
