+++ src/components/ManaCalculator/types/validator.type.ts
@@ -1,0 +1,14 @@
+export interface ValidatorProps {
+  lockedStake: number;
+  delegatedStake: number;
+  performanceFactor: number;
+  fixedCost: number;
+}
+
+export interface ValidatorParameters {
+  shareOfYourStakeLocked: number;
+  attractedNewDelegatedStake: number;
+  attractedDelegatedStakeFromOtherPools: number;
+  fixedCost: number;
+  performanceFactor: number;
+}
